refact: [NETWORK] split badquality views
