+++ FloconDesktop/composeApp/src/commonMain/kotlin/io/github/openflocon/flocondesktop/features/network/badquality/edition/view/BadQualityEditionWindow.kt
@@ -2,24 +2,13 @@
 
 package io.github.openflocon.flocondesktop.features.network.badquality.edition.view
 
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyRow
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Delete
 import androidx.compose.material3.BasicAlertDialog
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Text
@@ -28,19 +17,11 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.ColorFilter
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.util.fastForEach
 import io.github.openflocon.flocondesktop.features.network.badquality.edition.model.BadQualityConfigUiModel
 import io.github.openflocon.flocondesktop.features.network.badquality.edition.model.SelectedBadQualityUiModel
-import io.github.openflocon.flocondesktop.features.network.badquality.edition.model.possibleExceptions
 import io.github.openflocon.library.designsystem.FloconTheme
-import io.github.openflocon.library.designsystem.components.FloconButton
 import io.github.openflocon.library.designsystem.components.FloconDialog
 import io.github.openflocon.library.designsystem.components.FloconDialogButtons
 import io.github.openflocon.library.designsystem.components.FloconSurface
@@ -169,7 +150,7 @@ fun BadNetworkQualityEditionContent(
             suffix = { Text(text = "%") },
             containerColor = FloconTheme.colorPalette.panel,
         )
-        ErrorsListView(
+        BadQualityErrorsListView(
             modifier = Modifier.padding(top = 16.dp),
             errors = errors,
             onErrorsClicked = { error ->
@@ -188,7 +169,7 @@ fun BadNetworkQualityEditionContent(
                 FloconSurface {
                     when (val t = selectedError.type) {
                         is BadQualityConfigUiModel.Error.Type.Body -> {
-                            ErrorsEditor(
+                            BadQualityErrorsEditor(
                                 error = selectedError,
                                 httpType = t,
                                 onErrorsChange = { error ->
@@ -207,7 +188,7 @@ fun BadNetworkQualityEditionContent(
                         }
 
                         is BadQualityConfigUiModel.Error.Type.Exception -> {
-                            ErrorExceptionEditor(
+                            BadQuaityErrorExceptionEditor(
                                 error = selectedError,
                                 errorException = t,
                                 onErrorsChange = { error ->
@@ -267,255 +248,3 @@ fun BadNetworkQualityEditionContent(
     }
 }
 
-@Composable
-fun ErrorsListView(
-    errors: List<BadQualityConfigUiModel.Error>,
-    onErrorsClicked: (error: BadQualityConfigUiModel.Error) -> Unit,
-    deleteError: (error: BadQualityConfigUiModel.Error) -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Column(
-        modifier = modifier,
-        verticalArrangement = Arrangement.spacedBy(8.dp),
-    ) {
-        Row(
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(4.dp),
-        ) {
-            Text("Errors", style = FloconTheme.typography.titleMedium)
-            FloconButton(
-                onClick = {
-                    onErrorsClicked(
-                        BadQualityConfigUiModel.Error(
-                            // new error
-                            weight = 1f,
-                            type = BadQualityConfigUiModel.Error.Type.Body(
-                                httpCode = 500,
-                                body = "{\"error\":\"...\"}",
-                                contentType = "application/json",
-                            ),
-                        ),
-                    )
-                },
-            ) {
-                Text(
-                    text = "Add http error",
-                )
-            }
-            FloconButton(
-                onClick = {
-                    onErrorsClicked(
-                        BadQualityConfigUiModel.Error(
-                            // new error
-                            weight = 1f,
-                            type = BadQualityConfigUiModel.Error.Type.Exception(
-                                classPath = possibleExceptions.first().classPath,
-                            ),
-                        ),
-                    )
-                },
-            ) {
-                Text(
-                    text = "Add Exception",
-                )
-            }
-        }
-        LazyRow(
-            horizontalArrangement = Arrangement.spacedBy(4.dp),
-            modifier = Modifier.fillMaxWidth(),
-        ) {
-            items(errors) { error ->
-                Column(
-                    modifier = Modifier
-                        .width(230.dp)
-                        .clip(RoundedCornerShape(8.dp))
-                        .background(Color(0xFFEFEFEF).copy(alpha = 0.2f))
-                        .clickable {
-                            onErrorsClicked(error)
-                        }
-                        .padding(8.dp),
-                ) {
-                    val textStyle = FloconTheme.typography.bodyMedium.copy(
-                        fontWeight = FontWeight.Thin,
-                        color = FloconTheme.colorPalette.onSurface,
-                    )
-
-                    Text("Weight : ${error.weight}", style = textStyle)
-                    when (val t = error.type) {
-                        is BadQualityConfigUiModel.Error.Type.Body -> {
-                            Text("HttpCode : ${t.httpCode}", style = textStyle) // or throwable ?
-                            Text(t.contentType, style = textStyle)
-                            Text("Body : ${t.body}", maxLines = 2, style = textStyle)
-                        }
-
-                        is BadQualityConfigUiModel.Error.Type.Exception -> {
-                            Text(
-                                "Exception",
-                                style = textStyle
-                            )
-                            Text(
-                                t.classPath,
-                                style = textStyle
-                            )
-                        }
-                    }
-
-                    // bouton supprimer
-                    Row(
-                        modifier = Modifier.fillMaxWidth(),
-                        horizontalArrangement = Arrangement.End,
-                    ) {
-                        Image(
-                            imageVector = Icons.Default.Delete,
-                            contentDescription = "Delete error",
-                            modifier = Modifier
-                                .size(24.dp)
-                                .clickable {
-                                    deleteError(
-                                        error,
-                                    )
-                                },
-                            colorFilter = ColorFilter.tint(FloconTheme.colorPalette.onSurface),
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun ErrorExceptionEditor(
-    error: BadQualityConfigUiModel.Error,
-    errorException: BadQualityConfigUiModel.Error.Type.Exception,
-    onErrorsChange: (BadQualityConfigUiModel.Error) -> Unit,
-) {
-    Column(
-        modifier = Modifier.padding(all = 8.dp),
-        verticalArrangement = Arrangement.spacedBy(4.dp)
-    ) {
-        possibleExceptions.fastForEach { exception ->
-            val (backgroundColor, textColor) = if (exception.classPath == errorException.classPath) {
-                FloconTheme.colorPalette.onSurface to FloconTheme.colorPalette.panel
-            } else {
-                FloconTheme.colorPalette.panel to FloconTheme.colorPalette.onSurface
-            }
-            Column(
-                modifier = Modifier
-                    .clip(RoundedCornerShape(8.dp))
-                    .background(
-                        color = backgroundColor,
-                    )
-                    .then(
-                        Modifier.clickable(
-                            onClick = {
-                                onErrorsChange(
-                                    error.copy(
-                                        weight = 1f,
-                                        type = errorException.copy(
-                                            classPath = exception.classPath
-                                        )
-                                    )
-                                )
-                            },
-                        )
-                    ).padding(all = 8.dp)
-            ) {
-                Text(
-                    text = exception.description,
-                    style = FloconTheme.typography.bodySmall,
-                    color = textColor,
-                )
-                Text(
-                    text = exception.classPath,
-                    style = FloconTheme.typography.bodyMedium.copy(
-                        fontWeight = FontWeight.Bold,
-                    ),
-                    color = textColor,
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun ErrorsEditor(
-    error: BadQualityConfigUiModel.Error,
-    httpType: BadQualityConfigUiModel.Error.Type.Body,
-    onErrorsChange: (BadQualityConfigUiModel.Error) -> Unit,
-) {
-    var weight by remember(error) { mutableStateOf<String>(error.weight.toString()) }
-    var httpCode by remember(error) { mutableStateOf<String>(httpType.httpCode.toString()) }
-    var contentType by remember { mutableStateOf<String>(httpType.contentType) }
-    var body by remember(error) { mutableStateOf<String>(httpType.body) }
-
-    Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-        // bouton ajouter
-        Box(
-            modifier = Modifier
-                .clip(RoundedCornerShape(12.dp))
-                .background(FloconTheme.colorPalette.onSurface)
-                .clickable {
-                    onErrorsChange(
-                        error.copy(
-                            weight = weight.toFloatOrNull() ?: error.weight,
-                            type = httpType.copy(
-                                httpCode = httpCode.toIntOrNull() ?: httpType.httpCode,
-                                contentType = contentType,
-                                body = body,
-                            )
-                        ),
-                    )
-                }
-                .padding(horizontal = 8.dp, vertical = 4.dp),
-        ) {
-            Text(
-                "Save",
-                style = FloconTheme.typography.titleSmall,
-                color = FloconTheme.colorPalette.panel,
-            )
-        }
-        Column(
-            modifier = Modifier
-                .clip(RoundedCornerShape(8.dp))
-                .background(Color(0xFFEFEFEF).copy(alpha = 0.2f))
-                .padding(8.dp),
-        ) {
-            FloconTextField(
-                label = defaultLabel("Weight"),
-                placeholder = defaultPlaceHolder("eg: 1.0"),
-                value = weight,
-                onValueChange = {
-                    if (it.isEmpty() || it.toFloatOrNull() != null) {
-                        weight = it
-                    }
-                },
-            )
-
-            FloconTextField(
-                label = defaultLabel("HTTP Code"),
-                placeholder = defaultPlaceHolder("eg: 500"),
-                value = httpCode,
-                onValueChange = {
-                    if (it.isEmpty() || it.toIntOrNull() != null) {
-                        httpCode = it
-                    }
-                },
-            )
-
-            FloconTextField(
-                label = defaultLabel("Content-Type"),
-                placeholder = defaultPlaceHolder("application/json"),
-                value = contentType,
-                onValueChange = { contentType = it }
-            )
-
-            FloconTextField(
-                label = defaultLabel("Body"),
-                placeholder = defaultPlaceHolder("{\"error\":\"...\"}"),
-                value = body,
-                onValueChange = { body = it },
-            )
-        }
-    }
-}
